support easier local builds
According to [official GHA docs][1], free public runners only have 4
cores, 16GiB of RAM and ~14GiB of storage each, which is unlikely to
support a complete build of this repo within 6 hours. (As a comparison,
a complete build on my 3990X rig takes ~1hr to complete.)

Hence, provide a script that can be used to easily recreate the
(hopefully reasonably reproducible) artifacts on a local box. For now
the GHA YAML definitions are not parsed for less dependencies even
though introducing redundant info that needs to stay in sync. Hope this
will not be very burdensome to maintain...

[1]: https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories

Signed-off-by: WANG Xuerui <git@xen0n.name>